Added sovereign-forge structs to common structs module
