+++ sovereign-forge/src/common/utils.rs
@@ -3,45 +3,12 @@ use error_messages::{
     CHAIN_ID_NOT_FOUR_CHAR_LONG, CHAIN_ID_NOT_LOWERCASE_ALPHANUMERIC, DEPLOY_COST_NOT_ENOUGH,
     ESDT_SAFE_NOT_DEPLOYED, FEE_MARKET_NOT_DEPLOYED, HEADER_VERIFIER_NOT_DEPLOYED,
 };
-use multiversx_sc::{
-    api::ManagedTypeApi,
-    codec,
-    derive::{type_abi, ManagedVecItem},
-    proxy_imports::{NestedDecode, NestedEncode, TopDecode, TopEncode},
-    require,
-    types::ManagedAddress,
-};
+use multiversx_sc::require;
+use structs::forge::ScArray;
 
 const CHARSET: &[u8] = b"0123456789abcdefghijklmnopqrstuvwxyz";
 
 use crate::err_msg;
-
-#[type_abi]
-#[derive(TopEncode, TopDecode, NestedEncode, NestedDecode, ManagedVecItem)]
-pub struct ContractInfo<M: ManagedTypeApi> {
-    pub id: ScArray,
-    pub address: ManagedAddress<M>,
-}
-
-impl<M: ManagedTypeApi> ContractInfo<M> {
-    pub fn new(id: ScArray, address: ManagedAddress<M>) -> Self {
-        ContractInfo { id, address }
-    }
-}
-
-#[type_abi]
-#[derive(TopEncode, TopDecode, NestedEncode, NestedDecode, Clone, ManagedVecItem, PartialEq)]
-pub enum ScArray {
-    ChainFactory,
-    Controller,
-    HeaderVerifier,
-    ESDTSafe,
-    EnshrineESDTSafe,
-    FeeMarket,
-    TokenHandler,
-    ChainConfig,
-    Slashing,
-}
 
 const NUMBER_OF_SHARDS: u32 = 3;
